feat(ui): New Arrivals module with shared AlbumCard (KAMP-84)
- Expose added_at (MIN date_added) on AlbumInfo/AlbumOut so the frontend
  can filter by recency without a separate query
- Implement NewArrivalsModule: albums added in the last 30 days, sorted
  newest first, with loading skeleton and empty state
- Extract AlbumCard into a shared component so Base Kamp modules get
  context menu, drag-to-queue, and keyboard nav for free; AlbumGrid
  now imports the shared card
