+++ kamp_core/server.py
@@ -83,6 +83,8 @@ class AlbumOut(BaseModel):
     # MAX(file_mtime) across the album's tracks — appended to art URLs as ?v=
     # so the browser caches images by URL and only re-fetches when files change.
     art_version: float | None = None
+    # MIN(date_added) across the album's tracks — used by the New Arrivals module.
+    added_at: float | None = None
 
 
 class PlayerStateOut(BaseModel):
@@ -443,6 +445,7 @@ def create_app(
                 missing_album=a.missing_album,
                 file_path=a.file_path,
                 art_version=a.art_version,
+                added_at=a.added_at,
             )
             for a in index.albums(sort=sort)
         ]
@@ -527,6 +530,7 @@ def create_app(
                 missing_album=a.missing_album,
                 file_path=a.file_path,
                 art_version=a.art_version,
+                added_at=a.added_at,
             )
             for a in index.albums(sort=sort)
             if (a.album_artist, a.album) in fts_keys
